JS-158 Prepare generator to produce Java classes
