+++ tools/estree/package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "generate": "ts-node generate-proto-file.ts",
+    "generate-proto": "ts-node main.ts proto",
+    "generate-java": "ts-node main.ts java",
     "load": "ts-node load-proto.ts",
     "compile": "tsc"
   },
