--- tools/protobuf/load-proto.ts
@@ -1,3 +1,0 @@
-import protobuf from 'protobufjs';
-
-const proto = protobuf.loadSync('output/ast.proto');
